クラス構造を修正
